LayoutTests/platform/mac/media/audio-session-category-audio-autoplay-expected.txt


RUN(internals.settings.setShouldManageAudioSessionCategory(true))
RUN(audio.autoplay = true)
RUN(audio.src = findMediaFile("audio", "../../../media/content/test"))
EVENT(playing)
EXPECTED (internals.audioSessionCategory() == 'MediaPlayback') OK
END OF TEST
